try to fix qt6 build

+++ src/plugins/score-plugin-scenario/Scenario/Document/Interval/FullView/TimeSignatureItem.hpp
@@ -70,7 +70,7 @@ public:
       case Qt::Key_Right:
       {
         auto c = textCursor();
-        c.setPosition(ossia::min(toPlainText().size(), c.position() + 1));
+        c.setPosition(ossia::min(int64_t(toPlainText().size()), c.position() + 1));
         setTextCursor(c);
         return;
       }
